setup

+++ main.h
@@ -1,5 +1,5 @@
 #include "pico/stdlib.h"
-
+#include <stdint.h>
 float freqs[] = {27.5f,
 29.13524f,
 30.86771f,
@@ -88,4 +88,4 @@ float freqs[] = {27.5f,
 3729.31f,
 3951.066f,
 4186.009f};
-static const uint buzzerPin = 18;
+static const uint64_t buzzerPin = 18;
